refactor: workflow model 구조 정리
