+++ src/entities/workflow/model/useAddWorkflowNodeMutation.ts
@@ -1,7 +1,8 @@
 import { useMutation } from "@tanstack/react-query";
 
-import { type NodeAddRequest, workflowApi } from "@/entities/workflow";
 import { type MutationPolicyOptions, toMutationMeta } from "@/shared/api";
+
+import { type NodeAddRequest, workflowApi } from "../api";
 
 import { syncWorkflowCache } from "./workflow-cache-utils";
