+++ src/entities/workflow/model/useDeleteWorkflowMutation.ts
@@ -1,7 +1,8 @@
 import { useMutation } from "@tanstack/react-query";
 
-import { workflowApi } from "@/entities/workflow";
 import { type MutationPolicyOptions, toMutationMeta } from "@/shared/api";
+
+import { workflowApi } from "../api";
 
 import { removeWorkflowDomainCache } from "./workflow-cache-utils";
