+++ src/entities/workflow/model/useShareWorkflowMutation.ts
@@ -1,7 +1,8 @@
 import { useMutation } from "@tanstack/react-query";
 
-import { type ShareRequest, workflowApi } from "@/entities/workflow";
 import { type MutationPolicyOptions, toMutationMeta } from "@/shared/api";
+
+import { type ShareRequest, workflowApi } from "../api";
 
 import { invalidateWorkflowLists } from "./workflow-cache-utils";
