+++ src/entities/workflow/model/useWorkflowChoicesQuery.ts
@@ -1,12 +1,14 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { workflowApi } from "@/entities/workflow";
-import { workflowKeys } from "../../constants";
 import {
   type QueryPolicyOptions,
   resolveQueryPolicyOptions,
   toQueryMeta,
 } from "@/shared/api";
+
+import { workflowApi } from "../api";
+
+import { workflowKeys } from "./query-keys";
 
 export const useWorkflowChoicesQuery = (
   workflowId: string | undefined,
