+++ src/entities/workflow/model/useWorkflowListQuery.ts
@@ -1,12 +1,14 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { workflowApi } from "@/entities/workflow";
-import { workflowKeys } from "../../constants";
 import {
   type QueryPolicyOptions,
   resolveQueryPolicyOptions,
   toQueryMeta,
 } from "@/shared/api";
+
+import { workflowApi } from "../api";
+
+import { workflowKeys } from "./query-keys";
 
 export const useWorkflowListQuery = (
   page = 0,
