+++ src/features/add-node/ui/ServiceSelectionPanel.tsx
@@ -12,13 +12,15 @@ import { Box, Grid, Icon, Input, Text, VStack } from "@chakra-ui/react";
 import { useReactFlow, useViewport } from "@xyflow/react";
 
 import { NODE_REGISTRY } from "@/entities/node";
+import {
+  useAddWorkflowNodeMutation,
+  useDeleteWorkflowNodeMutation,
+} from "@/entities/workflow";
 import type { FlowNodeData, NodeMeta } from "@/entities/node";
 import {
   findAddedNodeId,
   toFlowNode,
   toNodeAddRequest,
-  useAddWorkflowNodeMutation,
-  useDeleteWorkflowNodeMutation,
   useWorkflowStore,
 } from "@/shared";
