+++ src/shared/model/index.ts
@@ -1,2 +1,1 @@
-export * from "./workflow";
 export * from "./workflowStore";
