+++ src/widgets/output-panel/ui/OutputPanel.tsx
@@ -5,6 +5,16 @@ import { Box, Icon, Spinner, Text, VStack } from "@chakra-ui/react";
 
 import { NODE_REGISTRY } from "@/entities/node";
 import type { DataType, NodeMeta, NodeType } from "@/entities/node";
+import {
+  type ChoiceBranchConfig,
+  type ChoiceFollowUp,
+  type ChoiceOption,
+  type ChoiceResponse,
+  useAddWorkflowNodeMutation,
+  useDeleteWorkflowNodeMutation,
+  useSelectWorkflowChoiceMutation,
+  useWorkflowChoicesQuery,
+} from "@/entities/workflow";
 import { useAddNode } from "@/features/add-node";
 import {
   MAPPING_NODE_TYPE_MAP,
@@ -24,18 +34,8 @@ import {
   findAddedNodeId,
   toFlowNode,
   toNodeAddRequest,
-  useAddWorkflowNodeMutation,
-  useDeleteWorkflowNodeMutation,
   useDualPanelLayout,
-  useSelectWorkflowChoiceMutation,
-  useWorkflowChoicesQuery,
   useWorkflowStore,
-} from "@/shared";
-import type {
-  ChoiceBranchConfig,
-  ChoiceFollowUp,
-  ChoiceOption,
-  ChoiceResponse,
 } from "@/shared";
 
 import {
